add basic cli options
