+++ Cargo.toml
@@ -2,9 +2,13 @@
 name = "words-rs"
 version = "0.1.0"
 edition = "2021"
+author = "Sarah Allen"
+description = "generate HTML with a span for each word"
+
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 anyhow = "1.0.81"
+clap = {version = "4.5.0", features = ["derive"]}
 regex = "1.10.3"
